Don't raise error for legit tmp files
This commit fixes an error when chaining multiple filters. Filters will
correctly put files inside the tmp directory. This commit verifies that
all files are not inside the tmp directory when input_files is
calculated. This will raise an error when you match files inside the
pipeline's tmp dir or specifically add it as an input